requirements.txt
flask
psycopg2-binary
configparser
pandas